Image popper
Takes two inputs when started, search terms and 'seconds', and then
checks google images with your search terms, and opens a window with
your default browser every X number of seconds, with a random image
based on said search terms.

Makes a file named 'akito.xml' in the same directory to store the data
it retrieves image links from.

Does not check if the image was deleted, so you may occasionally get a
blank page, or a page with any site's "Image has been removed" message.